pass on the code and readme
