+++ bfv/plaintext.go
@@ -45,6 +45,9 @@ func NewPlaintextLvl(params Parameters, level int) *Plaintext {
 // where the message is set to the passed poly. No checks are performed on poly and
 // the returned Plaintext will share its backing array of coefficient.
 func NewPlaintextAtLevelFromPoly(level int, poly *ring.Poly) *Plaintext {
+	if len(poly.Coeffs) < level+1 {
+		panic("provided ring.Poly level is too small")
+	}
 	v0 := new(ring.Poly)
 	v0.Coeffs = poly.Coeffs[:level+1]
 	return &Plaintext{Plaintext: &rlwe.Plaintext{Value: v0}}
